GB-56 feat : admin, user login 기능 구현

+++ src/main/java/garlicbears/quiz/domain/management/user/controller/UserController.java
@@ -1,11 +1,16 @@
 package garlicbears.quiz.domain.management.user.controller;
 
+import java.util.List;
 import java.util.Map;
 import java.util.logging.Logger;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,15 +21,17 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import garlicbears.quiz.domain.common.dto.ResponseDto;
+import garlicbears.quiz.domain.common.entity.Role;
 import garlicbears.quiz.domain.common.entity.User;
+import garlicbears.quiz.domain.management.common.dto.LoginDto;
 import garlicbears.quiz.domain.management.common.dto.ResponseUserDto;
 import garlicbears.quiz.domain.management.user.dto.SignUpDto;
 import garlicbears.quiz.domain.management.user.dto.UpdateUserDto;
 import garlicbears.quiz.domain.management.user.service.UserRatingService;
 import garlicbears.quiz.domain.management.user.service.UserService;
-import garlicbears.quiz.global.config.auth.PrincipalDetails;
 import garlicbears.quiz.global.exception.CustomException;
 import garlicbears.quiz.global.exception.ErrorCode;
+import garlicbears.quiz.global.jwt.util.JwtTokenizer;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 
@@ -35,12 +42,18 @@ public class UserController implements SwaggerUserController {
 	private static final Logger logger = Logger.getLogger(UserController.class.getName());
 	private final UserService userService;
 	private final UserRatingService userRatingService;
+	private final JwtTokenizer jwtTokenizer;
+	private final PasswordEncoder passwordEncoder;
 
 	@Autowired
 	UserController(UserService userService,
-		UserRatingService userRatingService) {
+		UserRatingService userRatingService,
+		JwtTokenizer jwtTokenizer,
+		PasswordEncoder passwordEncoder) {
 		this.userService = userService;
 		this.userRatingService = userRatingService;
+		this.jwtTokenizer = jwtTokenizer;
+		this.passwordEncoder = passwordEncoder;
 	}
 
 	/**
@@ -93,9 +106,13 @@ public class UserController implements SwaggerUserController {
 	}
 
 	@GetMapping("/")
-	public ResponseEntity<?> searchUser(@AuthenticationPrincipal PrincipalDetails principalDetails) {
+	public ResponseEntity<?> searchUser(@AuthenticationPrincipal UserDetails userDetails) {
 		// 현재 인증된 사용자의 정보를 principalDetails로부터 가져올 수 있습니다.
-		User user = principalDetails.getUser();
+		if (userDetails == null) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
+		}
+		// 현재 인증된 사용자의 정보를 UserDetails로부터 가져올 수 있습니다.
+		User user = userService.findByEmail(userDetails.getUsername());
 
 		System.out.println(user);
 
@@ -103,10 +120,13 @@ public class UserController implements SwaggerUserController {
 	}
 
 	@PatchMapping("/")
-	public ResponseEntity<?> updateUser(@AuthenticationPrincipal PrincipalDetails principalDetails,
+	public ResponseEntity<?> updateUser(@AuthenticationPrincipal UserDetails userDetails,
 		@RequestBody UpdateUserDto updateUserDto) {
-		// 현재 인증된 사용자의 정보를 principalDetails로부터 가져올 수 있습니다.
-		User user = principalDetails.getUser();
+		if (userDetails == null) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
+		}
+		// 현재 인증된 사용자의 정보를 UserDetails로부터 가져올 수 있습니다.
+		User user = userService.findByEmail(userDetails.getUsername());
 
 		userService.update(user, updateUserDto);
 
@@ -114,8 +134,12 @@ public class UserController implements SwaggerUserController {
 	}
 
 	@DeleteMapping("/")
-	public ResponseEntity<?> deleteUser(@AuthenticationPrincipal PrincipalDetails principalDetails) {
-		User user = principalDetails.getUser();
+	public ResponseEntity<?> deleteUser(@AuthenticationPrincipal UserDetails userDetails) {
+		if (userDetails == null) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
+		}
+		// 현재 인증된 사용자의 정보를 UserDetails로부터 가져올 수 있습니다.
+		User user = userService.findByEmail(userDetails.getUsername());
 
 		userService.delete(user);
 
@@ -123,13 +147,45 @@ public class UserController implements SwaggerUserController {
 	}
 
 	@PostMapping("/rating")
-	public ResponseEntity<?> rating(@AuthenticationPrincipal PrincipalDetails principalDetails,
+	public ResponseEntity<?> rating(@AuthenticationPrincipal UserDetails userDetails,
 		@RequestBody Map<String, Double> request) {
-		User user = principalDetails.getUser();
+		if (userDetails == null) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
+		}
+		// 현재 인증된 사용자의 정보를 UserDetails로부터 가져올 수 있습니다.
+		User user = userService.findByEmail(userDetails.getUsername());
 		Double ratingValue = request.get("ratingValue");
 
 		userRatingService.saveRating(user, ratingValue);
 
 		return ResponseEntity.ok(ResponseDto.success());
 	}
+
+	@PostMapping("/login")
+	public ResponseEntity<?> login(@Valid @RequestBody LoginDto loginDto, BindingResult bindingResult) {
+		bindingResult.getAllErrors().forEach(error -> logger.warning(error.toString()));
+		if (bindingResult.hasErrors()) {
+
+			throw new CustomException(ErrorCode.INVALID_INPUT);
+		}
+
+		User user = userService.findByEmail(loginDto.getEmail());
+		if (!passwordEncoder.matches(loginDto.getPassword(), user.getPassword())) {
+			throw new CustomException(ErrorCode.INVALID_INPUT);
+		}
+
+		List<String> roles = user.getRoles().stream()
+			.map(Role::getRoleName)
+			.toList();
+
+		String accessToken = jwtTokenizer.createAccessToken(user.getUserEmail(), user.getUserId(), roles);
+		String refreshToken = jwtTokenizer.createRefreshToken(user.getUserEmail(), user.getUserId(), roles);
+
+		HttpHeaders headers = new HttpHeaders();
+		headers.set("Access-Token", accessToken);
+		headers.set("Refresh-Token", refreshToken);
+
+		return ResponseEntity.ok().headers(headers).body(ResponseDto.success());
+	}
+
 }
